+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Command line flags: `--help`, `--version`
+- `--config-path`/`-c` command line flag to set an alternate config path for
+  one execution
 
 ## [0.1.0] - 2022-07-03
 ### Added